+++ src/Rotation/Rotation.cpp
@@ -2,7 +2,7 @@
 
 namespace devuino
 {
-    void Rotation::radian(float rotate)
+    void Rotation::radian(double rotate)
     {
         degree(rotate * (180.0 / 3.1415));
     }
